[REV] project, project_* : revert the refactoring of task.template
This commit's purpose is to revert the refactor of the task template
done in task-4919945. There should be no functional changes due to
this commit. The revert is done because the refactor adds too many
code lines and is too complex.

Task-5072878
