+++ addons/project/models/__init__.py
@@ -1,7 +1,6 @@
 # -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
-from . import project_task_template
 from . import account_analytic_account
 from . import mail_message
 from . import project_project_stage
