+++ addons/project/models/project_milestone.py
@@ -6,7 +6,7 @@ from collections import defaultdict
 from odoo import api, fields, models
 from odoo.tools import format_date
 
-from .project_task_template import CLOSED_STATES
+from .project_task import CLOSED_STATES
 
 
 class ProjectMilestone(models.Model):
