+++ addons/project/models/project_task.py
@@ -1,14 +1,14 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
-from .project_task_template import CLOSED_STATES
-
+import re
 from pytz import UTC
 from collections import defaultdict
 from datetime import timedelta, datetime, time
 
-from odoo import api, fields, models, tools, _
+from odoo import api, fields, models, tools, SUPERUSER_ID, _
 from odoo.fields import Command, Date, Domain
 from odoo.addons.rating.models import rating_data
+from odoo.addons.web_editor.tools import handle_history_divergence
 from odoo.exceptions import UserError, ValidationError
 from odoo.tools import format_list, SQL, LazyTranslate
 from odoo.addons.resource.models.utils import filter_domain_leaf
@@ -54,6 +54,8 @@ PROJECT_TASK_READABLE_FIELDS = {
     'recurring_count',
     'duration_tracking',
     'display_follow_button',
+    'is_template',
+    'has_template_ancestor',
     'stage_id_color',
     'access_token',
     'access_url',
@@ -76,23 +78,33 @@ PROJECT_TASK_WRITABLE_FIELDS = {
     'is_closed',
 }
 
+CLOSED_STATES = {
+    '1_done': 'Done',
+    '1_canceled': 'Cancelled',
+}
+
 
 class ProjectTask(models.Model):
     _name = 'project.task'
     _description = "Task"
     _date_name = "date_assign"
     _inherit = [
-        'project.task.template',
         'portal.mixin',
         'mail.thread.cc',
         'mail.activity.mixin',
         'rating.mixin',
         'mail.tracking.duration.mixin',
+        'html.field.history.mixin',
     ]
     _mail_post_access = 'read'
     _mail_thread_customer = True
+    _order = "priority desc, sequence, date_deadline asc, id desc"
     _primary_email = 'email_from'
+    _systray_view = 'list'
     _track_duration_field = 'stage_id'
+
+    def _get_versioned_fields(self):
+        return [ProjectTask.description.name]
 
     @api.model
     def _get_default_partner_id(self, project=None, parent=None):
@@ -102,24 +114,93 @@ class ProjectTask(models.Model):
             return project.partner_id.id
         return False
 
+    def _get_default_stage_id(self):
+        """ Gives default stage_id """
+        project_id = self.env.context.get('default_project_id')
+        if not project_id:
+            return False
+        return self.stage_find(project_id, order="fold, sequence, id")
+
     @api.model
     def _default_user_ids(self):
         return self.env.user.ids if any(key in self.env.context for key in ('default_personal_stage_type_ids', 'default_personal_stage_type_id')) else ()
 
     @api.model
+    def _default_company_id(self):
+        if self.env.context.get('default_project_id'):
+            return self.env['project.project'].browse(self.env.context['default_project_id']).company_id
+        return False
+
+    @api.model
+    def _read_group_stage_ids(self, stages, domain):
+        search_domain = [('id', 'in', stages.ids)]
+        if 'default_project_id' in self.env.context and not self.env.context.get(
+                'subtask_action') and 'project_kanban' in self.env.context:
+            search_domain = ['|', ('project_ids', '=', self.env.context['default_project_id'])] + search_domain
+
+        stage_ids = stages._search(search_domain, order=stages._order)
+        return stages.browse(stage_ids)
+
+    @api.model
     def _read_group_personal_stage_type_ids(self, stages, domain):
         return stages.search(['|', ('id', 'in', stages.ids), ('user_id', '=', self.env.user.id)])
 
-    name = fields.Char(tracking=True)
-    stage_id = fields.Many2one(tracking=True)
-    state = fields.Selection(tracking=True, inverse='_inverse_state')
-    priority = fields.Selection(tracking=True)
+    active = fields.Boolean(default=True, export_string_translation=False)
+    name = fields.Char(string='Title', tracking=True, required=True, index='trigram')
+    description = fields.Html(string='Description', sanitize_attributes=False)
+    priority = fields.Selection([
+        ('0', 'Low priority'),
+        ('1', 'Medium priority'),
+        ('2', 'High priority'),
+        ('3', 'Urgent'),
+    ], default='0', index=True, string="Priority", tracking=True)
+    sequence = fields.Integer(string='Sequence', default=10, export_string_translation=False)
+    stage_id = fields.Many2one('project.task.type', string='Stage', compute='_compute_stage_id',
+       store=True, readonly=False, ondelete='restrict', tracking=True, index=True,
+       default=_get_default_stage_id, group_expand='_read_group_stage_ids',
+       domain="[('project_ids', '=', project_id)]")
+    stage_id_color = fields.Integer(string='Stage Color', related="stage_id.color", export_string_translation=False)
+    tag_ids = fields.Many2many('project.tags', string='Tags')
+
+    state = fields.Selection([
+        ('01_in_progress', 'In Progress'),
+        ('02_changes_requested', 'Changes Requested'),
+        ('03_approved', 'Approved'),
+        *CLOSED_STATES.items(),
+        ('04_waiting_normal', 'Waiting'),
+    ], string='State', copy=False, default='01_in_progress', required=True, compute='_compute_state',
+        inverse='_inverse_state', readonly=False, store=True, index=True, recursive=True, tracking=True)
+    is_closed = fields.Boolean("Closed state", compute='_compute_is_closed', search='_search_is_closed')
+
+    create_date = fields.Datetime("Created On", readonly=True, index=True)
+    write_date = fields.Datetime("Last Updated On", readonly=True)
+    date_end = fields.Datetime(string='Ending Date', index=True, copy=False)
     date_assign = fields.Datetime(string='Assigning Date', copy=False, readonly=True,
         help="Date on which this task was last assigned (or unassigned). Based on this, you can get statistics on the time it usually takes to assign tasks.")
-    date_deadline = fields.Datetime(tracking=True, copy=False)
-    project_id = fields.Many2one('project.project', required=False, falsy_value_label=_lt("🔒 Private"))
-    allocated_hours = fields.Float(tracking=True)
-    user_ids = fields.Many2many('res.users', relation='project_task_user_rel', column1='task_id', column2='user_id', string='Assignees', context={'active_test': False}, tracking=True, default=_default_user_ids, domain="[('share', '=', False), ('active', '=', True)]", falsy_value_label=_lt("👤 Unassigned"))
+    date_deadline = fields.Datetime(string='Deadline', index=True, tracking=True, copy=False)
+
+    date_last_stage_update = fields.Datetime(string='Last Stage Update',
+        index=True,
+        copy=False,
+        readonly=True,
+        help="Date on which the state of your task has last been modified.\n"
+            "Based on this information you can identify tasks that are stalling and get statistics on the time it usually takes to move tasks from one stage/state to another.")
+
+    project_id = fields.Many2one('project.project', string='Project', domain="['|', ('company_id', '=', False), ('company_id', '=?',  company_id), ('is_template', 'in', [is_template, False])]",
+                                 compute="_compute_project_id", store=True, precompute=True, recursive=True, readonly=False, index=True, tracking=True, change_default=True, falsy_value_label=_lt("🔒 Private"))
+    display_in_project = fields.Boolean(compute='_compute_display_in_project', store=True, export_string_translation=False)
+    task_properties = fields.Properties('Properties', definition='project_id.task_properties_definition', copy=True)
+    allocated_hours = fields.Float("Allocated Time", tracking=True)
+    subtask_allocated_hours = fields.Float("Sub-tasks Allocated Time", compute='_compute_subtask_allocated_hours', export_string_translation=False,
+        help="Sum of the hours allocated for all the sub-tasks (and their own sub-tasks) linked to this task. Usually less than or equal to the allocated hours of this task.")
+    role_ids = fields.Many2many(
+        'project.role',
+        string='Project Roles',
+        help="When you create a project from a template, you can choose which employee takes each role. These employees will be added to the tasks, along with anyone already assigned.",
+    )
+    # Tracking of this field is done in the write function
+    user_ids = fields.Many2many('res.users', relation='project_task_user_rel', column1='task_id', column2='user_id',
+        string='Assignees', context={'active_test': False}, tracking=True, default=_default_user_ids, domain="[('share', '=', False), ('active', '=', True)]", falsy_value_label=_lt("👤 Unassigned"))
     # User names displayed in project sharing views
     portal_user_names = fields.Char(compute='_compute_portal_user_names', compute_sudo=True, search='_search_portal_user_names', export_string_translation=False)
     # Second Many2many containing the actual personal stage for the current user
@@ -148,11 +229,25 @@ class ProjectTask(models.Model):
     # Need this field to check there is no email loops when Odoo reply automatically
     email_from = fields.Char('Email From')
     email_cc = fields.Char(help='Email addresses that were in the CC of the incoming emails from this task and that are not currently linked to an existing customer.')
+    company_id = fields.Many2one('res.company', string='Company', compute='_compute_company_id', store=True, readonly=False, recursive=True, copy=True, default=_default_company_id)
+    color = fields.Integer(string='Color Index', export_string_translation=False)
     rating_active = fields.Boolean(string='Project Rating Status', related="project_id.rating_active")
+    attachment_ids = fields.One2many(
+        'ir.attachment',
+        compute='_compute_attachment_ids',
+        string="Attachments",
+        export_string_translation=False,
+        help="Attachments that don't come from a message",
+    )
+    # In the domain of displayed_image_id, we couln't use attachment_ids because a one2many is represented as a list of commands so we used res_model & res_id
+    displayed_image_id = fields.Many2one('ir.attachment', domain="[('res_model', '=', 'project.task'), ('res_id', '=', id), ('mimetype', 'ilike', 'image')]", string='Cover Image')
 
     parent_id = fields.Many2one('project.task', string='Parent Task', inverse="_inverse_parent_id", index=True, domain="['!', ('id', 'child_of', id)]", tracking=True)
     child_ids = fields.One2many('project.task', 'parent_id', string="Sub-tasks", domain="[('recurring_task', '=', False)]", export_string_translation=False)
+    subtask_count = fields.Integer("Sub-task Count", compute='_compute_subtask_count', export_string_translation=False)
+    closed_subtask_count = fields.Integer("Closed Sub-tasks Count", compute='_compute_subtask_count', export_string_translation=False)
     project_privacy_visibility = fields.Selection(related='project_id.privacy_visibility', string="Project Visibility", tracking=False)
+    subtask_completion_percentage = fields.Float(compute="_compute_subtask_completion_percentage", export_string_translation=False)
     # Computed field about working time elapsed between record creation and assignation/closing.
     working_hours_open = fields.Float(compute='_compute_elapsed', string='Working Hours to Assign', digits=(16, 2), store=True, aggregator="avg")
     working_hours_close = fields.Float(compute='_compute_elapsed', string='Working Hours to Close', digits=(16, 2), store=True, aggregator="avg")
@@ -160,13 +255,25 @@ class ProjectTask(models.Model):
     working_days_close = fields.Float(compute='_compute_elapsed', string='Working Days to Close', store=True, aggregator="avg")
     # customer portal: include comment and (incoming/outgoing) emails in communication history
     website_message_ids = fields.One2many(domain=lambda self: [('model', '=', self._name), ('message_type', 'in', ['email', 'comment', 'email_outgoing', 'auto_comment'])], export_string_translation=False)
-    milestone_id = fields.Many2one(tracking=True)
-    task_template_id = fields.Many2one(
-        comodel_name='project.task.template',
-        string='Task Template',
-        copy=False,
+    allow_milestones = fields.Boolean(related='project_id.allow_milestones', export_string_translation=False)
+    milestone_id = fields.Many2one(
+        'project.milestone',
+        'Milestone',
+        domain="[('project_id', '=', project_id)]",
+        compute='_compute_milestone_id',
+        readonly=False,
+        store=True,
+        tracking=True,
+        index='btree_not_null',
+        help="Deliver your services automatically when a milestone is reached by linking it to a sales order item."
     )
-
+    has_late_and_unreached_milestone = fields.Boolean(
+        compute='_compute_has_late_and_unreached_milestone',
+        search='_search_has_late_and_unreached_milestone',
+        export_string_translation=False,
+    )
+    # Task Dependencies fields
+    allow_task_dependencies = fields.Boolean(related='project_id.allow_task_dependencies', export_string_translation=False)
     # Tracking of this field is done in the write function
     depend_on_ids = fields.Many2many('project.task', relation="task_dependencies_rel", column1="task_id",
                                      column2="depends_on_id", string="Blocked By", tracking=True, copy=False,
@@ -184,16 +291,38 @@ class ProjectTask(models.Model):
     display_follow_button = fields.Boolean(compute='_compute_display_follow_button', compute_sudo=True, export_string_translation=False)
 
     # recurrence fields
+    recurring_task = fields.Boolean(string="Recurrent")
     recurring_count = fields.Integer(string="Tasks in Recurrence", compute='_compute_recurring_count')
     recurrence_id = fields.Many2one('project.task.recurrence', copy=False, index='btree_not_null')
-    repeat_interval = fields.Integer(compute='_compute_repeat', compute_sudo=True, readonly=False, default=1)
-    repeat_unit = fields.Selection(compute='_compute_repeat', compute_sudo=True, readonly=False)
-    repeat_type = fields.Selection(
-        selection_add=[('until', 'Until')],
-        default="forever", string="Until", compute='_compute_repeat', compute_sudo=True, readonly=False)
-    repeat_until = fields.Date(compute='_compute_repeat', compute_sudo=True, readonly=False)
+    repeat_interval = fields.Integer(string='Repeat Every', default=1, compute='_compute_repeat', compute_sudo=True, readonly=False)
+    repeat_unit = fields.Selection([
+        ('day', 'Days'),
+        ('week', 'Weeks'),
+        ('month', 'Months'),
+        ('year', 'Years'),
+    ], default='week', compute='_compute_repeat', compute_sudo=True, readonly=False)
+    repeat_type = fields.Selection([
+        ('forever', 'Forever'),
+        ('until', 'Until'),
+    ], default="forever", string="Until", compute='_compute_repeat', compute_sudo=True, readonly=False)
+    repeat_until = fields.Date(string="End Date", compute='_compute_repeat', compute_sudo=True, readonly=False)
 
+    # Quick creation shortcuts
+    display_name = fields.Char(
+        inverse='_inverse_display_name',
+        help="""Use these keywords in the title to set new tasks:\n
+            #tags Set tags on the task
+            @user Assign the task to a user
+            ! Set the task a medium priority
+            !! Set the task a high priority
+            !!! Set the task a urgent priority\n
+            Make sure to use the right format and order e.g. Improve the configuration screen #feature #v16 @Mitchell !""",
+    )
     link_preview_name = fields.Char(compute='_compute_link_preview_name', export_string_translation=False)
+    is_template = fields.Boolean(copy=False, export_string_translation=False)
+    has_project_template = fields.Boolean(related='project_id.is_template', string="Has Project Template", export_string_translation=False)
+    has_template_ancestor = fields.Boolean(compute='_compute_has_template_ancestor', search='_search_has_template_ancestor',
+                                           recursive=True, export_string_translation=False)
 
     _recurring_task_has_no_parent = models.Constraint(
         'CHECK (NOT (recurring_task IS TRUE AND parent_id IS NOT NULL))',
@@ -203,6 +332,8 @@ class ProjectTask(models.Model):
         'CHECK (NOT (project_id IS NULL AND parent_id IS NOT NULL))',
         'A private task cannot have a parent.',
     )
+
+    _is_template_idx = models.Index('(is_template) WHERE is_template IS TRUE')
 
     @api.constrains('company_id', 'partner_id')
     def _ensure_company_consistency_with_partner(self):
@@ -226,11 +357,72 @@ class ProjectTask(models.Model):
     def TASK_PORTAL_WRITABLE_FIELDS(self):
         return PROJECT_TASK_WRITABLE_FIELDS
 
+    @api.depends('parent_id.project_id')
+    def _compute_project_id(self):
+        self.env.remove_to_compute(self._fields['display_in_project'], self)
+        for task in self:
+            if not task.display_in_project and task.parent_id and task.parent_id.project_id != task.project_id:
+                task.project_id = task.parent_id.project_id
+
+    @api.depends('project_id', 'parent_id')
+    def _compute_display_in_project(self):
+        for record in self:
+            record.display_in_project = not record.project_id or (
+                    not record.parent_id or record.project_id != record.parent_id.project_id
+            )
+
+    def _inverse_parent_id(self):
+        for task in self.sudo():
+            if not task.parent_id:
+                task.display_in_project = True
+            elif task.display_in_project and task.project_id == task.parent_id.sudo().project_id:
+                task.display_in_project = False
+
+    @api.depends('stage_id', 'depend_on_ids.state')
+    def _compute_state(self):
+        for task in self:
+            dependent_open_tasks = []
+            if task.allow_task_dependencies:
+                dependent_open_tasks = [dependent_task for dependent_task in task.depend_on_ids if
+                                        dependent_task.state not in CLOSED_STATES]
+            # if one of the blocking task is in a blocking state
+            if dependent_open_tasks:
+                # here we check that the blocked task is not already in a closed state (if the task is already done we don't put it in waiting state)
+                if task.state not in CLOSED_STATES:
+                    task.state = '04_waiting_normal'
+            # if the task as no blocking dependencies and is in waiting_normal, the task goes back to in progress
+            elif task.state not in CLOSED_STATES:
+                task.state = '01_in_progress'
+
+    @api.depends('state')
+    def _compute_is_closed(self):
+        for task in self:
+            task.is_closed = task.state in CLOSED_STATES
+
+    def _search_is_closed(self, operator, value):
+        if operator == 'in':
+            searched_states = list(CLOSED_STATES.keys())
+        elif operator == 'not in':
+            searched_states = self.OPEN_STATES
+        else:
+            return NotImplemented
+        return [('state', 'in', searched_states)]
+
     def _get_rotting_depends_fields(self):
         return super()._get_rotting_depends_fields() + ['is_closed']
 
     def _get_rotting_domain(self):
         return super()._get_rotting_domain() & Domain('is_closed', '=', False)
+
+    @property
+    def OPEN_STATES(self):
+        """ Return a list of the technical names complementing the CLOSED_STATES, a.k.a the open states """
+        return list(set(self._fields['state'].get_values(self.env)) - set(CLOSED_STATES))
+
+    @api.onchange('project_id')
+    def _onchange_project_id(self):
+        if self.state != '04_waiting_normal':
+            self.state = '01_in_progress'
 
     def is_blocked_by_dependences(self):
         return any(blocking_task.state not in CLOSED_STATES for blocking_task in self.depend_on_ids)
@@ -306,7 +498,12 @@ class ProjectTask(models.Model):
 
     @api.model
     def _get_recurrence_fields(self):
-        return super()._get_recurrence_fields() + ['repeat_until']
+        return [
+            'repeat_interval',
+            'repeat_unit',
+            'repeat_type',
+            'repeat_until',
+        ]
 
     @api.depends('recurring_task')
     def _compute_repeat(self):
@@ -376,6 +573,21 @@ class ProjectTask(models.Model):
             for task in tasks_with_dependency:
                 task.dependent_tasks_count = dependent_tasks_count_dict.get(task.id, 0)
 
+    @api.constrains('parent_id')
+    def _check_parent_id(self):
+        if self._has_cycle():
+            raise ValidationError(_('Error! You cannot create a recursive hierarchy of tasks.'))
+
+    def _get_attachments_search_domain(self):
+        self.ensure_one()
+        return [('res_id', '=', self.id), ('res_model', '=', 'project.task')]
+
+    def _compute_attachment_ids(self):
+        for task in self:
+            attachment_ids = self.env['ir.attachment'].search(task._get_attachments_search_domain()).ids
+            message_attachment_ids = task.mapped('message_ids.attachment_ids').ids  # from mail_thread
+            task.attachment_ids = [(6, 0, list(set(attachment_ids) - set(message_attachment_ids)))]
+
     @api.depends('create_date', 'date_end', 'date_assign')
     def _compute_elapsed(self):
         task_linked_to_calendar = self.filtered(
@@ -410,6 +622,28 @@ class ProjectTask(models.Model):
         for task in self:
             task.access_url = f'/my/tasks/{task.id}'
 
+    @api.depends('child_ids.allocated_hours')
+    def _compute_subtask_allocated_hours(self):
+        for task in self:
+            task.subtask_allocated_hours = sum(task.child_ids.mapped('allocated_hours'))
+
+    @api.depends('child_ids')
+    def _compute_subtask_count(self):
+        if not any(self._ids):
+            for task in self:
+                task.subtask_count, task.closed_subtask_count = len(task.child_ids), len(task.child_ids.filtered(lambda r: r.state in CLOSED_STATES))
+            return
+        total_and_closed_subtask_count_per_parent_id = {
+            parent.id: (count, sum(s in CLOSED_STATES for s in states))
+            for parent, states, count in self.env['project.task']._read_group(
+                [('parent_id', 'in', self.ids)],
+                ['parent_id'],
+                ['state:array_agg', '__count'],
+            )
+        }
+        for task in self:
+            task.subtask_count, task.closed_subtask_count = total_and_closed_subtask_count_per_parent_id.get(task.id, (0, 0))
+
     @api.depends('partner_id.phone')
     def _compute_partner_phone(self):
         for task in self:
@@ -419,6 +653,28 @@ class ProjectTask(models.Model):
         for task in self:
             if task.partner_id:
                 task.partner_id.phone = task.partner_phone
+
+    @api.onchange('company_id')
+    def _onchange_task_company(self):
+        if self.project_id.company_id and self.project_id.company_id != self.company_id:
+            self.project_id = False
+
+    @api.depends('project_id.company_id', 'parent_id.company_id')
+    def _compute_company_id(self):
+        for task in self:
+            if not task.parent_id and not task.project_id:
+                continue
+            task.company_id = task.project_id.company_id or task.parent_id.company_id
+
+    @api.depends('project_id')
+    def _compute_stage_id(self):
+        for task in self:
+            project = task.project_id or task.parent_id.project_id
+            if project:
+                if project not in task.stage_id.project_ids:
+                    task.stage_id = task.stage_find(project.id, [('fold', '=', False)])
+            else:
+                task.stage_id = False
 
     @api.depends('user_ids')
     def _compute_portal_user_names(self):
@@ -474,12 +730,97 @@ class ProjectTask(models.Model):
         for task in self:
             task.display_follow_button = not limited_access_per_project_id.get(task.project_id, True)
 
+    def _get_group_pattern(self):
+        return {
+            'tags_and_users': r'\s([#@]%s[^\s]+)',
+            'priority': r'(?:^|\s)(!{1,3})(?=\s|$)',
+        }
+
+    def _prepare_pattern_groups(self):
+        group = self._get_group_pattern()
+        return [
+            group['tags_and_users'] % '',
+            group['priority'],
+        ]
+
+    def _get_groups_patterns(self):
+        return [
+            r'(?:%s)*' % ('|').join(self._prepare_pattern_groups()),
+        ]
+
+    def _get_cannot_start_with_patterns(self):
+        return [r'(?![#!@\s])']
+
+    def _extract_tags_and_users(self):
+        tags = []
+        users = []
+        tags_and_users_group = self._get_group_pattern()['tags_and_users']
+        for word in re.findall(tags_and_users_group % '', self.display_name):
+            (tags if word.startswith('#') else users).append(word[1:])
+        users_to_keep = []
+        user_ids = []
+        for user in users:
+            matched_users = self.env['res.users'].name_search(user)
+            if len(matched_users) == 1:
+                user_ids.append(Command.link(matched_users[0][0]))
+            else:
+                users_to_keep.append(r'%s\b' % user)
+        self.user_ids = user_ids
+        if tags:
+            domain = Domain.OR(Domain('name', '=ilike', tag) for tag in tags)
+            existing_tags = self.env['project.tags'].search(domain)
+            existing_tags_names = {tag.name.lower() for tag in existing_tags}
+            new_tags_names = {tag for tag in tags if tag.lower() not in existing_tags_names}
+            self.tag_ids = [Command.set(existing_tags.ids)] + [Command.create({'name': name}) for name in new_tags_names]
+        pattern = tags_and_users_group % ('(?!%s)' % ('|').join(users_to_keep) if users_to_keep else '')
+        self.display_name, _ = re.subn(pattern, '', self.display_name)
+
+    def _extract_priority(self):
+        priority_group = self._get_group_pattern()['priority']
+        match = re.search(priority_group, self.display_name)
+        if match:
+            self.priority = str(min(len(match.group(1)), 3))
+            self.display_name, _dummy = re.subn(priority_group, '', self.display_name)
+
+    def _get_groups(self):
+        return [
+            lambda task: task._extract_tags_and_users(),
+            lambda task: task._extract_priority(),
+        ]
+
+    def _inverse_display_name(self):
+        for task in self:
+            pattern = re.compile(r'^%s.+?%s$' % (
+                ('').join(task._get_cannot_start_with_patterns()),
+                ('').join(task._get_groups_patterns()))
+            )
+            match = pattern.match(task.display_name)
+            if match:
+                for group, extract_data in enumerate(task._get_groups(), start=1):
+                    if match.group(group):
+                        extract_data(task)
+                task.name = task.display_name.strip()
+
     def _compute_link_preview_name(self):
         for task in self:
             link_preview_name = task.display_name
             if task.project_id:
                 link_preview_name += f' | {task.project_id.sudo().name}'
             task.link_preview_name = link_preview_name
+
+    @api.depends('is_template', 'parent_id.has_template_ancestor')
+    def _compute_has_template_ancestor(self):
+        for task in self:
+            task.has_template_ancestor = task.is_template or (task.parent_id and task.parent_id.sudo().has_template_ancestor)
+
+    def _search_has_template_ancestor(self, operator, value):
+        if operator not in ['=', '!='] or not isinstance(value, bool):
+            return NotImplemented
+        template_tasks = self.env['project.task'].with_context(active_test=False).sudo().search([('is_template', '=', True)])
+        domain = [('id', 'child_of', template_tasks.ids)]
+        if (operator == "=") != value:
+            domain = ['!', ('id', 'child_of', template_tasks.ids)]
+        return domain
 
     def copy_data(self, default=None):
         default = dict(default or {})
@@ -498,9 +839,40 @@ class ProjectTask(models.Model):
             for vals in vals_list
         ]
 
+        active_users = self.env['res.users']
+        has_default_users = 'user_ids' in default
+        if not has_default_users:
+            active_users = self.user_ids.filtered('active')
+        milestone_mapping = self.env.context.get('milestone_mapping', {})
         for task, vals in zip(self, vals_list):
+
+            if not default.get('stage_id'):
+                vals['stage_id'] = task.stage_id.id
+            if 'active' not in default and not task['active'] and not self.env.context.get('copy_project'):
+                vals['active'] = True
+            if not default.get('name'):
+                vals['name'] = task.name if self.env.context.get('copy_project') or self.env.context.get('copy_from_template') else _("%s (copy)", task.name)
             if task.recurrence_id and not default.get('recurrence_id'):
                 vals['recurrence_id'] = task.recurrence_id.copy().id
+            if task.allow_milestones:
+                vals['milestone_id'] = milestone_mapping.get(vals['milestone_id'], vals['milestone_id'])
+            if not default.get('child_ids') and task.child_ids:
+                default = {
+                    'parent_id': False,
+                }
+                current_task = task
+                if self.env.context.get('copy_from_template'):
+                    current_task = current_task.with_context(active_test=True)
+                child_ids = current_task.child_ids
+                vals['child_ids'] = [Command.create(child_id.copy_data(default)[0]) for child_id in child_ids]
+            if not has_default_users and vals['user_ids']:
+                task_active_users = task.user_ids & active_users
+                vals['user_ids'] = [Command.set(task_active_users.ids)]
+            if task.is_template and not self.env.context.get('copy_from_template'):
+                vals['is_template'] = True
+            if self.env.context.get('copy_from_template'):
+                for field in set(self._get_template_field_blacklist()) & set(vals.keys()):
+                    del vals[field]
         return vals_list
 
     def _create_task_mapping(self, copied_tasks):
@@ -571,6 +943,31 @@ class ProjectTask(models.Model):
         )
         return super().get_empty_list_help(help_message)
 
+    # ----------------------------------------
+    # Case management
+    # ----------------------------------------
+
+    def stage_find(self, section_id, domain=[], order='sequence, id'):
+        """ Override of the base.stage method
+            Parameter of the stage search taken from the lead:
+            - section_id: if set, stages must belong to this section or
+              be a default stage; if not set, stages must be default
+              stages
+        """
+        # collect all section_ids
+        section_ids = []
+        if section_id:
+            section_ids.append(section_id)
+        section_ids.extend(self.mapped('project_id').ids)
+        search_domain = []
+        if section_ids:
+            search_domain = [('|')] * (len(section_ids) - 1)
+            for section_id in section_ids:
+                search_domain.append(('project_ids', '=', section_id))
+        search_domain += list(domain)
+        # perform search, return the first found
+        return self.env['project.task.type'].search(search_domain, order=order, limit=1).id
+
     # ------------------------------------------------
     # CRUD overrides
     # ------------------------------------------------
@@ -585,6 +982,13 @@ class ProjectTask(models.Model):
     @api.model
     def default_get(self, fields):
         vals = super().default_get(fields)
+
+        if project_id := self.env.context.get('default_create_in_project_id'):
+            vals['project_id'] = project_id
+
+        # prevent creating new task in the waiting state
+        if 'state' in fields and vals.get('state') == '04_waiting_normal':
+            vals['state'] = '01_in_progress'
 
         if 'repeat_until' in fields:
             vals['repeat_until'] = Date.today() + timedelta(days=7)
@@ -609,6 +1013,13 @@ class ProjectTask(models.Model):
             user_ids = vals.get('user_ids', [])
             user_ids.append(Command.link(self.env.user.id))
             vals['user_ids'] = user_ids
+
+        parent_id = vals.get('parent_id', self.env.context.get('default_parent_id'))
+        if parent_id:
+            parent = self.env['project.task'].browse(parent_id)
+            if not vals.get('tag_ids'):
+                vals['tag_ids'] = parent.tag_ids
+
         return vals
 
     @api.model
@@ -643,6 +1054,15 @@ class ProjectTask(models.Model):
             if field and field.type == 'many2one':
                 self.env[field.comodel_name].browse(value).check_access('read')
 
+    def _set_stage_on_project_from_task(self):
+        stage_ids_per_project = defaultdict(list)
+        for task in self:
+            if task.stage_id and task.stage_id not in task.project_id.type_ids and task.stage_id.id not in stage_ids_per_project[task.project_id]:
+                stage_ids_per_project[task.project_id].append(task.stage_id.id)
+
+        for project, stage_ids in stage_ids_per_project.items():
+            project.write({'type_ids': [Command.link(stage_id) for stage_id in stage_ids]})
+
     def _load_records_create(self, vals_list):
         for vals in vals_list:
             if vals.get('recurring_task'):
@@ -665,26 +1085,80 @@ class ProjectTask(models.Model):
 
         new_context = dict(self.env.context)
         default_personal_stage = new_context.pop('default_personal_stage_type_ids', False)
+        default_project_id = new_context.pop('default_project_id', False)
+        if not default_project_id:
+            parent_task = self.browse({parent_id for vals in vals_list if (parent_id := vals.get('parent_id'))})
+            if len(parent_task) == 1:
+                default_project_id = parent_task.sudo().project_id.id
+        # (portal) users that don't have write access can still create a task
+        # in the project that will be checked using record rules
+        new_context["default_create_in_project_id"] = default_project_id
+        if not self._has_field_access(self._fields['user_ids'], 'write'):
+            # remove user_ids if we have no access to it
+            new_context.pop('default_user_ids', False)
+        self = self.with_context(new_context)
+
+        self.browse().check_access('create')
+        default_stage = dict()
         for vals, additional_vals in zip(vals_list, additional_vals_list):
+            project_id = vals.get('project_id') or default_project_id
+
+            if vals.get('user_ids'):
+                additional_vals['date_assign'] = fields.Datetime.now()
+                if not (vals.get('parent_id') or project_id):
+                    user_ids = self._fields['user_ids'].convert_to_cache(vals.get('user_ids', []), self.env['project.task'])
+                    if self.env.user.id not in list(user_ids) + [SUPERUSER_ID]:
+                        additional_vals['user_ids'] = [Command.set(list(user_ids) + [self.env.user.id])]
             if default_personal_stage and 'personal_stage_type_id' not in vals:
                 additional_vals['personal_stage_type_id'] = default_personal_stage[0]
-        # create the task, write computed inaccessible fields in sudo
-        for vals, computed_vals in zip(vals_list, additional_vals_list):
+            if not vals.get('name') and vals.get('display_name'):
+                vals['name'] = vals['display_name']
+
             if self.env.user._is_portal() and not self.env.su:
                 self._ensure_fields_write(vals, defaults=True)
+
+            if project_id and not "company_id" in vals:
+                additional_vals["company_id"] = self.env["project.project"].browse(
+                    project_id
+                ).company_id.id
+            if not project_id and ("stage_id" in vals or self.env.context.get('default_stage_id')):
+                vals["stage_id"] = False
+
+            if project_id and "stage_id" not in vals:
+                # 1) Allows keeping the batch creation of tasks
+                # 2) Ensure the defaults are correct (and computed once by project),
+                # by using default get (instead of _get_default_stage_id or _stage_find),
+                if project_id not in default_stage:
+                    default_stage[project_id] = self.with_context(
+                        default_project_id=project_id
+                    ).default_get(['stage_id']).get('stage_id')
+                vals["stage_id"] = default_stage[project_id]
+
+            # Stage change: Update date_end if folded stage and date_last_stage_update
+            if vals.get('stage_id'):
+                additional_vals.update(self.update_date_end(vals['stage_id']))
+                additional_vals['date_last_stage_update'] = fields.Datetime.now()
             # recurrence
             rec_fields = vals.keys() & self._get_recurrence_fields()
             if rec_fields and vals.get('recurring_task') is True:
                 rec_values = {rec_field: vals[rec_field] for rec_field in rec_fields}
                 recurrence = self.env['project.task.recurrence'].create(rec_values)
                 vals['recurrence_id'] = recurrence.id
+
+        # create the task, write computed inaccessible fields in sudo
+        for vals, computed_vals in zip(vals_list, additional_vals_list):
             for field_name in list(computed_vals):
                 if self._has_field_access(self._fields[field_name], 'write'):
                     vals[field_name] = computed_vals.pop(field_name)
-        tasks = super().create(vals_list)
+        # no track when the portal user create a task to avoid using during tracking
+        # process since the portal does not have access to tracking models
+        tasks = super(ProjectTask, self.with_context(mail_create_nosubscribe=True, mail_notrack=not self.env.su and self.env.user._is_portal())).create(vals_list)
+        for task, computed_vals in zip(tasks.sudo(), additional_vals_list):
+            if computed_vals:
+                task.write(computed_vals)
+        tasks.sudo()._populate_missing_personal_stages()
         self._task_message_auto_subscribe_notify({task: task.user_ids - self.env.user for task in tasks})
 
-        tasks._populate_missing_personal_stages()
         current_partner = self.env.user.partner_id
 
         all_partner_emails = []
@@ -696,6 +1170,8 @@ class ProjectTask(models.Model):
             for partner in partners
             if not all(u.share for u in partner.user_ids)
         }
+        if tasks.project_id:
+            tasks.sudo()._set_stage_on_project_from_task()
         for task in tasks.sudo():
             if task.project_id.privacy_visibility in ['invited_users', 'portal']:
                 task._portal_ensure_token()
@@ -717,13 +1193,69 @@ class ProjectTask(models.Model):
 
     def write(self, vals):
         self.check_access('write')
+        if len(self) == 1:
+            handle_history_divergence(self, 'description', vals)
         partner_ids = []
+
+        # Some values are determined by this override and must be written as
+        # sudo for portal users, because they do not have access to these
+        # fields. Other values must not be written as sudo.
+        additional_vals = {}
         if self.env.user._is_portal() and not self.env.su:
             self._ensure_fields_write(vals, defaults=False)
 
+        if 'milestone_id' in vals:
+            # WARNING: has to be done after 'project_id' vals is written on subtasks
+            milestone = self.env['project.milestone'].browse(vals['milestone_id'])
+
+            # 1. Task for which the milestone is unvalid -> milestone_id is reset
+            if 'project_id' not in vals:
+                unvalid_milestone_tasks = self.filtered(lambda task: task.project_id != milestone.project_id) if vals['milestone_id'] else self.env['project.task']
+            else:
+                unvalid_milestone_tasks = self if not vals['milestone_id'] or milestone.project_id.id != vals['project_id'] else self.env['project.task']
+            valid_milestone_tasks = self - unvalid_milestone_tasks
+            if unvalid_milestone_tasks:
+                unvalid_milestone_tasks.sudo().write({'milestone_id': False})
+                if valid_milestone_tasks:
+                    valid_milestone_tasks.sudo().write({'milestone_id': vals['milestone_id']})
+                del vals['milestone_id']
+
+            # 2. Parent's milestone is set to subtask with no milestone recursively
+            subtasks_to_update = valid_milestone_tasks.child_ids.filtered(
+                lambda task: (task not in self and
+                              not task.milestone_id and
+                              task.project_id == milestone.project_id and
+                              task.state not in CLOSED_STATES))
+
+            # 3. If parent and child task share the same milestone, child task's milestone is updated when the parent one is changed
+            # No need to check if state is changed in vals as it won't affect the subtasks selected for update
+            if 'project_id' not in vals:
+                subtasks_to_update |= valid_milestone_tasks.child_ids.filtered(
+                    lambda task: (task not in self and
+                                  task.milestone_id == task.parent_id.milestone_id and
+                                  task.state not in CLOSED_STATES))
+            else:
+                subtasks_to_update |= valid_milestone_tasks.child_ids.filtered(
+                    lambda task: (task not in self and
+                                  (not task.display_in_project or task.project_id.id == vals['project_id']) and
+                                  task.milestone_id == task.parent_id.milestone_id and
+                                  task.state not in CLOSED_STATES))
+            if subtasks_to_update:
+                subtasks_to_update.sudo().write({'milestone_id': vals['milestone_id']})
+
+        if vals.get('parent_id') in self.ids:
+            raise UserError(_("Sorry. You can't set a task as its parent task."))
+
+        # stage change: update date_last_stage_update
         now = fields.Datetime.now()
-        # task_ids_without_user_set = set()
-        if 'user_ids' in vals:
+        if 'stage_id' in vals:
+            if not 'project_id' in vals and self.filtered(lambda t: not t.project_id):
+                raise UserError(_('You can only set a personal stage on a private task.'))
+
+            additional_vals.update(self.update_date_end(vals['stage_id']))
+            additional_vals['date_last_stage_update'] = now
+        task_ids_without_user_set = set()
+        if 'user_ids' in vals and 'date_assign' not in vals:
             # prepare update of date_assign after super call
             task_ids_without_user_set = {task.id for task in self if not task.user_ids}
 
@@ -764,6 +1296,18 @@ class ProjectTask(models.Model):
                         if not project_link:
                             project_link = link_per_project_id[task.project_id.id] = task.project_id._get_html_link(title=task.project_id.display_name)
                         project_link_per_task_id[task.id] = project_link
+        if vals.get('parent_id') is False:
+            additional_vals['display_in_project'] = True
+        if 'description' in vals:
+            # the portal user cannot access to html_field_history and so it would be
+            # better to write in sudo for description field to avoid giving access to html_field_history
+            additional_vals['description'] = vals.pop('description')
+
+            # write changes
+        if self.env.su or not self.env.user._is_portal():
+            vals.update(additional_vals)
+        elif additional_vals:
+            super(ProjectTask, self.sudo()).write(additional_vals)
         result = super().write(vals)
 
         if 'user_ids' in vals:
@@ -788,6 +1332,12 @@ class ProjectTask(models.Model):
                     if task.is_blocked_by_dependences() and vals['state'] not in CLOSED_STATES and vals['state'] != '04_waiting_normal':
                         task.state = '04_waiting_normal'
                 task.date_last_stage_update = now
+        elif 'project_id' in vals:
+            self.filtered(lambda t: t.state != '04_waiting_normal').state = '01_in_progress'
+
+        # Do not recompute the state when changing the parent (to avoid resetting the state)
+        if 'parent_id' in vals:
+            self.env.remove_to_compute(self._fields['state'], self)
 
         self._task_message_auto_subscribe_notify({task: task.user_ids - old_user_ids[task] - self.env.user for task in self})
 
@@ -818,6 +1368,12 @@ class ProjectTask(models.Model):
             if task.id == last_task_id_per_recurrence_id.get(task.recurrence_id.id):
                 task.recurrence_id.unlink()
         return super().unlink()
+
+    def update_date_end(self, stage_id):
+        project_task_type = self.env['project.task.type'].browse(stage_id)
+        if project_task_type.fold:
+            return {'date_end': fields.Datetime.now()}
+        return {'date_end': False}
 
     def _search_on_comodel(self, domain, field, comodel, additional_domain=None):
         """ This method is called by `group_expand` methods, whose purpose is to add empty groups to the `read_group`
@@ -876,11 +1432,42 @@ class ProjectTask(models.Model):
             Use the project partner_id if any, or else the parent task partner_id.
         """
         for task in self:
+            if task.has_template_ancestor:
+                continue
             if task.partner_id and not (task.project_id or task.parent_id):
                 task.partner_id = False
                 continue
             if not task.partner_id:
                 task.partner_id = self._get_default_partner_id(task.project_id, task.parent_id)
+
+    @api.depends('project_id')
+    def _compute_milestone_id(self):
+        for task in self:
+            if task.project_id != task.milestone_id.project_id:
+                task.milestone_id = task.parent_id.project_id == task.project_id and task.parent_id.milestone_id
+
+    def _compute_has_late_and_unreached_milestone(self):
+        if all(not task.allow_milestones for task in self):
+            self.has_late_and_unreached_milestone = False
+            return
+        late_milestones = self.env['project.milestone'].sudo()._search([  # sudo is needed for the portal user in Project Sharing.
+            ('id', 'in', self.milestone_id.ids),
+            ('is_reached', '=', False),
+            ('deadline', '<=', fields.Date.today()),
+        ])
+        for task in self:
+            task.has_late_and_unreached_milestone = task.allow_milestones and task.milestone_id.id in late_milestones
+
+    def _search_has_late_and_unreached_milestone(self, operator, value):
+        if operator != 'in':
+            return NotImplemented
+        return [
+            ('allow_milestones', '=', True),
+            ('milestone_id', 'any', [
+                ('is_reached', '=', False),
+                ('deadline', '<', fields.Date.today()),
+            ]),
+        ]
 
     # ---------------------------------------------------
     # Mail gateway
@@ -980,7 +1567,7 @@ class ProjectTask(models.Model):
     def _track_template(self, changes):
         res = super()._track_template(changes)
         test_task = self[0]
-        if 'stage_id' in changes and test_task.stage_id.mail_template_id:
+        if 'stage_id' in changes and test_task.stage_id.mail_template_id and not test_task.is_template:
             res['stage_id'] = (test_task.stage_id.mail_template_id, {
                 'auto_delete_keep_log': False,
                 'subtype_id': self.env['ir.model.data']._xmlid_to_res_id('mail.mt_note'),
@@ -1166,6 +1753,63 @@ class ProjectTask(models.Model):
     def _get_projects_to_make_billable_domain(self, additional_domain=None):
         return Domain('partner_id', '!=', False) & Domain(additional_domain or Domain.TRUE)
 
+    def _get_all_subtasks(self):
+        return self.browse(set.union(set(), *self._get_subtask_ids_per_task_id().values()))
+
+    def _get_subtask_ids_per_task_id(self):
+        if not self:
+            return {}
+
+        res = dict.fromkeys(self._ids, [])
+        if all(self._ids):
+            self.env.cr.execute(
+                """
+         WITH RECURSIVE task_tree
+                     AS (
+                     SELECT id, id as supertask_id
+                       FROM project_task
+                      WHERE id IN %(ancestor_ids)s
+                      UNION
+                         SELECT t.id, tree.supertask_id
+                           FROM project_task t
+                           JOIN task_tree tree
+                             ON tree.id = t.parent_id
+                            AND t.active in (TRUE, %(active)s)
+                          WHERE t.parent_id IS NOT NULL
+               ) SELECT supertask_id, ARRAY_AGG(id)
+                   FROM task_tree
+                  WHERE id != supertask_id
+               GROUP BY supertask_id
+                """,
+                {
+                    "ancestor_ids": tuple(self.ids),
+                    "active": self.env.context.get('active_test', True),
+                }
+            )
+            res.update(dict(self.env.cr.fetchall()))
+        else:
+            res.update({
+                task.id: task._get_subtasks_recursively().ids
+                for task in self
+            })
+        return res
+
+    def _get_subtasks_recursively(self):
+        children = self.child_ids
+        if not children:
+            return self.env['project.task']
+        return children + children._get_subtasks_recursively()
+
+    def action_open_parent_task(self):
+        return {
+            'name': _('Parent Task'),
+            'view_mode': 'form',
+            'res_model': 'project.task',
+            'res_id': self.parent_id.id,
+            'type': 'ir.actions.act_window',
+            'context': self.env.context
+        }
+
     def action_project_sharing_view_parent_task(self):
         if self.parent_id.project_id != self.project_id and self.env.user._is_portal():
             project = self.parent_id.project_id._filtered_access('read')
@@ -1310,45 +1954,53 @@ class ProjectTask(models.Model):
             }
         }
 
-    def _project_task_template_vals(self):
-        defaults = {}
-        if self.recurring_task:
-            defaults = {
-                'repeat_interval': self.repeat_interval,
-                'repeat_unit': self.repeat_unit,
-                'repeat_type': self.repeat_type,
-                'repeat_until': self.repeat_until,
-            }
-        template_fields = self.env['project.task.template']._fields
-        generated_task = self.with_context(copy_project=True).copy_data(defaults)[0]
-
-        def filter_vals(vals):
-            """Recursively filter values by template_fields, including nested child_ids."""
-            filtered = {}
-            for f_name, value in vals.items():
-                if f_name in template_fields:
-                    if f_name == "child_ids" and value:
-                        filtered_children = []
-                        for command, _, child_vals in value:
-                            filtered_children.append((command, 0, filter_vals(child_vals)))
-                        filtered[f_name] = filtered_children
-                    else:
-                        filtered[f_name] = value
-            return filtered
-        return filter_vals(generated_task)
-
-    def action_generate_task_template(self):
+    def action_convert_to_template(self):
         self.ensure_one()
-        task_template_vals = self._project_task_template_vals()
-        task_template = self.env['project.task.template'].create(task_template_vals)
-        self.task_template_id = task_template.id
+        if not self.project_id:
+            return {
+                'type': 'ir.actions.client',
+                'tag': 'display_notification',
+                'params': {
+                    'type': 'danger',
+                    'message': _('Private tasks cannot be converted into templates'),
+                },
+            }
+        if self.is_template:
+            if self.project_id.is_template:
+                raise UserError(self.env._("Tasks in a project template cannot be converted into regular tasks."))
+
+            return {
+                'type': 'ir.actions.client',
+                'tag': 'project_show_template_undo_confirmation_dialog',
+                'params': {
+                    'task_id': self.id,
+                },
+            }
+        self.is_template = True
+        self.message_post(body=_("Task converted to template"))
         return {
             'type': 'ir.actions.client',
             'tag': 'project_show_template_notification',
             'params': {
-                'res_model': 'project.task.template',
-                'res_id': self.task_template_id.id,
-                'undo_method': "unlink",
+                'task_id': self.id,
+                'next': {
+                    'type': 'ir.actions.client',
+                    'tag': 'soft_reload',
+                },
+            },
+        }
+
+    def action_undo_convert_to_template(self):
+        self.ensure_one()
+        self.is_template = False
+        self.role_ids = False
+        self.message_post(body=_("Template converted back to regular task"))
+        return {
+            'type': 'ir.actions.client',
+            'tag': 'display_notification',
+            'params': {
+                'type': 'success',
+                'message': _('Template converted back to regular task'),
                 'next': {
                     'type': 'ir.actions.client',
                     'tag': 'soft_reload',
@@ -1360,6 +2012,44 @@ class ProjectTask(models.Model):
         self.ensure_one()
         return self.write(vals)
 
+    @api.model
+    def _get_template_default_context_whitelist(self):
+        """
+        Whitelist of fields that can be set through the `default_` context keys when creating a task from a template.
+        """
+        return [
+            "parent_id",
+        ]
+
+    @api.model
+    def _get_template_field_blacklist(self):
+        """
+        Blacklist of fields to not copy when creating a task from a template.
+        """
+        return [
+            "partner_id",
+        ]
+
+    def action_create_from_template(self, values=None):
+        self.ensure_one()
+        values = values or {}
+        default = {
+                      key[8:]: value
+                      for key, value in self.env.context.items()
+                      if key.startswith('default_') and key[8:] in self._get_template_default_context_whitelist()
+                  } | {
+                      field: False
+                      for field in self._get_template_field_blacklist()
+                  } | values
+        return self.with_context(copy_from_template=True).copy(default=default).id
+
+    def action_archive(self):
+        child_tasks = self.child_ids.filtered(lambda child_task: not child_task.display_in_project)
+        if child_tasks:
+            child_tasks.action_archive()
+        self.filtered(lambda t: not t.display_in_project and t.parent_id).display_in_project = True
+        return super().action_archive()
+
     # ---------------------------------------------------
     # Rating business
     # ---------------------------------------------------
@@ -1368,7 +2058,7 @@ class ProjectTask(models.Model):
         for task in self:
             rating_template = task.stage_id.rating_template_id
             partner = task.partner_id
-            if rating_template and partner and partner != self.env.user.partner_id:
+            if rating_template and partner and partner != self.env.user.partner_id and not task.is_template:
                 task.rating_send_request(rating_template, lang=task.partner_id.lang, force_send=force_send)
 
     def _rating_get_partner(self):
@@ -1446,6 +2136,11 @@ class ProjectTask(models.Model):
         else:
             self.sudo().message_subscribe(self.env.user.partner_id.ids)
         return not is_follower
+
+    @api.depends('subtask_count', 'closed_subtask_count')
+    def _compute_subtask_completion_percentage(self):
+        for task in self:
+            task.subtask_completion_percentage = task.subtask_count and task.closed_subtask_count / task.subtask_count
 
     @api.model
     def _get_allowed_access_params(self):
