+++ addons/project/security/ir.model.access.csv
@@ -8,7 +8,6 @@ access_project_task_type_project_user,project.task.type.project.user,model_proje
 access_project_task_type_manager,project.task.type manager,model_project_task_type,project.group_project_manager,1,1,1,1
 access_project_task_type_portal,task_type_portal,project.model_project_task_type,base.group_portal,1,0,0,0
 access_project_task,project.task,model_project_task,project.group_project_user,1,1,1,1
-access_project_task_template,project.task.template,model_project_task_template,project.group_project_user,1,1,1,1
 access_report_project_task_user,report.project.task.user,model_report_project_task_user,project.group_project_manager,1,0,0,0
 access_report_project_task_user_project_user,report.project.task.user.project.user,model_report_project_task_user,project.group_project_user,1,0,0,0
 access_partner_task_user,base.res.partner user,base.model_res_partner,project.group_project_user,1,0,0,0
