+++ addons/project/security/project_security.xml
@@ -81,11 +81,6 @@
         <field name="model_id" ref="model_project_task"/>
         <field name="domain_force">[('company_id', 'in', company_ids + [False])]</field>
     </record>
-    <record model="ir.rule" id="task_template_comp_rule">
-        <field name="name">Project/Task Template: multi-company</field>
-        <field name="model_id" ref="model_project_task_template"/>
-        <field name="domain_force">[('company_id', 'in', company_ids + [False])]</field>
-    </record>
 
     <record model="ir.rule" id="task_visibility_rule">
         <field name="name">Project/Task: employees: follow required for follower-only projects</field>
@@ -116,16 +111,6 @@
                  ('user_ids', 'in', user.id),
         ]</field>
         <field name="groups" eval="[(4,ref('project.group_project_manager'))]"/>
-    </record>
-
-    <record model="ir.rule" id="project_manager_all_task_template_rule">
-        <field name="name">Project/Task Template: project manager: see all task template linked to a project or its own templates</field>
-        <field name="model_id" ref="model_project_task"/>
-        <field name="domain_force">[
-            '|', ('project_id', '!=', False),
-                 ('user_ids', 'in', user.id),
-        ]</field>
-        <field name="groups" eval="[(4, ref('project.group_project_manager'))]"/>
     </record>
 
     <record model="ir.rule" id="task_type_manager_rule">
