+++ addons/project/static/src/actions/client_actions.js
@@ -2,15 +2,6 @@ import { ConfirmationDialog } from "@web/core/confirmation_dialog/confirmation_d
 import { registry } from "@web/core/registry";
 import { _t } from "@web/core/l10n/translation";
 
-export async function openFormView(env, { resModel, resId }) {
-    await env.services.action.doAction({
-        type: "ir.actions.act_window",
-        res_model: resModel,
-        views: [[false, "form"]],
-        res_id: resId,
-    });
-}
-
 export function showTemplateUndoNotification(
     env,
     {
@@ -20,8 +11,6 @@ export function showTemplateUndoNotification(
         undoMethod = "action_undo_convert_to_template",
         actionType = "success",
         undoCallback,
-        templateResId,
-        templateResModel,
     }
 ) {
     const undoNotification = env.services.notification.add(_t(message), {
@@ -38,20 +27,10 @@ export function showTemplateUndoNotification(
                     if (res && undoMethod !== "unlink") {
                         env.services.action.doAction(res);
                     } else if (undoMethod === "unlink") {
-                        if (model === 'project.task.template') {
-                            env.services.notification.add(_t('Removed from templates'), {
-                                type: "success",
-                            });
-                            env.services.action.doAction({
-                                type: "ir.actions.client",
-                                tag: "soft_reload",
-                            });
-                        } else {
-                            // Taking out the controller to be restored after unlinking the record
-                            const restoreController =
-                                env.services.action.currentController.config.breadcrumbs?.at(-2);
-                            restoreController?.onSelected();
-                        }
+                        // Taking out the controller to be restored after unlinking the record
+                        const restoreController =
+                            env.services.action.currentController.config.breadcrumbs?.at(-2);
+                        restoreController?.onSelected();
                     }
                     undoNotification();
                 },
@@ -77,11 +56,6 @@ export function showTemplateUndoConfirmationDialog(
         confirm: async () => {
             const action = await env.services.orm.call(model, undoMethod, [recordId]);
             await env.services.action.doAction(action);
-            if (action.params.type == 'success') {
-                const { res_id, res_model } = action.params;
-                await openFormView(env, { resModel: res_model, resId: res_id });
-                env.services.action.currentController.config.breadcrumbs?.pop()
-            }
             if (confirmationCallback) {
                 await env.services.orm.call(
                     model,
@@ -100,7 +74,12 @@ export async function showTemplateFormView(
     { model, recordId, method = "action_create_template_from_project" }
 ) {
     const action = await env.services.orm.call(model, method, [recordId]);
-    await openFormView(env, { resModel: model, resId: action.params.project_id });
+    await env.services.action.doAction({
+        type: "ir.actions.act_window",
+        res_model: model,
+        views: [[false, "form"]],
+        res_id: action.params.project_id,
+    });
     await env.services.action.doAction(action);
 }
 
@@ -108,13 +87,28 @@ export async function showTemplateFormView(
 registry.category("actions").add("project_show_template_notification", (env, action) => {
     const params = action.params || {};
     showTemplateUndoNotification(env, {
-        model: "project.task.template",
-        recordId: params.res_id,
-        undoMethod: params.undo_method,
-        message: _t("Saved as template"),
+        model: "project.task",
+        recordId: params.task_id,
+        message: _t("Task converted to template"),
     });
     return params.next;
 });
+
+// Task → Template Undo Confirmation Dialog
+registry
+    .category("actions")
+    .add("project_show_template_undo_confirmation_dialog", (env, action) => {
+        const params = action.params || {};
+        showTemplateUndoConfirmationDialog(env, {
+            model: "project.task",
+            recordId: params.task_id,
+            bodyMessage: _t(
+                "This task is currently a template. Would you like to convert it back into a regular task?"
+            ),
+            confirmLabel: _t("Convert to Task"),
+        });
+        return params.next;
+    });
 
 // Project → Template Create Redirection
 registry.category("actions").add("project_to_template_redirection_action", (env, action) => {
