+++ addons/project/static/src/views/components/project_task_template_dropdown.js
@@ -63,7 +63,7 @@ export class ProjectTaskTemplateDropdown extends Component {
         }
         this.action.switchView("form", {
             resId: await this.orm.call(
-                "project.task.template",
+                "project.task",
                 "action_create_from_template",
                 [templateId],
                 {
