+++ addons/project/views/project_milestone_views.xml
@@ -14,7 +14,6 @@
                                 context="{'default_project_id': project_id}"
                                 groups="project.group_project_milestone"
                                 close="1"
-                                invisible="not task_count"
                         >
                             <div class="o_form_field o_stat_info">
                                 <span class="o_stat_value">
