+++ addons/project_sms/models/project_task.py
@@ -9,7 +9,7 @@ class ProjectTask(models.Model):
 
     def _send_sms(self):
         for task in self:
-            if task.partner_id and task.stage_id and task.stage_id.sms_template_id:
+            if task.partner_id and task.stage_id and task.stage_id.sms_template_id and not task.is_template:
                 task._message_sms_with_template(
                     template=task.stage_id.sms_template_id,
                     partner_ids=task.partner_id.ids,
