+++ addons/sale_project/models/product_template.py
@@ -37,8 +37,8 @@ class ProductTemplate(models.Model):
         'project.project', 'Project Template', company_dependent=True, copy=True,
         domain='[("is_template", "=", True)]',
     )
-    task_template_id = fields.Many2one('project.task.template', 'Task Template',
-        domain="[('project_id', '=', project_id)]",
+    task_template_id = fields.Many2one('project.task', 'Task Template',
+        domain="[('is_template', '=', True), ('project_id', '=', project_id)]",
         company_dependent=True, copy=True, compute='_compute_task_template', store=True, readonly=False
     )
     service_policy = fields.Selection('_selection_service_policy', string="Service Invoicing Policy", compute_sudo=True, compute='_compute_service_policy', inverse='_inverse_service_policy', tracking=True)
